feat: footer에 help 버튼 추가, 갤러리 페이지에서 이젤 클릭시 help의 모달이 띄워지도록 변경
