+++ client/src/MainPage/MainWorld.tsx
@@ -6,7 +6,6 @@ import ErrorBoundary from "../components/common/ErrorBoundary";
 import Player from "./mapObjects/Player";
 import MapDataFetcher from "./components/MapDataFetcher";
 import Ground from "./mapObjects/Ground";
-import Environments from "./mapObjects/Environments";
 // import DevTools from "../components/Devtools";
 
 export default function MainWorld() {
@@ -19,7 +18,6 @@ export default function MainWorld() {
       </ErrorBoundary>
       <Light />
       <Ground />
-      <Environments />
       <Player />
       {/*<DevTools showDevtool={true} speed={10} />*/}
     </Physics>
