+++ client/src/components/Footer/helper.scss
@@ -63,14 +63,17 @@ button:focus-visible {
   .sign-buttons {
     display: flex;
     justify-content: space-between;
+    margin: 20px 0px;
     .sign-html-button {
       width: 250px;
       height: 30px;
       font-size: 15px;
       color: #ffffff;
       background-color: #2c2c2c;
-      margin: 20px 0px;
       border-radius: 10px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
     }
   }
 }
