+++ client/src/components/Footer/index.tsx
@@ -1,11 +1,18 @@
 import AudioHandler from "./AudioHandler";
 import ExpandButton from "./ExpandButton";
+import Helper from "./Helper";
 
 import "./style.scss";
 import GithubMark from "../../assets/images/github.svg";
 import ScreenShotIcon from "../../assets/images/screenshot.svg";
 
-export default function Footer() {
+import type { ReactNode } from "react";
+
+interface FooterProps {
+  helper: ReactNode;
+}
+
+export default function Footer({ helper }: FooterProps) {
   return (
     <div className="footer">
       <span className="footer-text">© Monument Gallery</span>
@@ -14,6 +21,7 @@ export default function Footer() {
           <img width={24} height={24} src={GithubMark} alt="github" />
         </button>
       </a>
+      {helper && <Helper>{helper}</Helper>}
       <AudioHandler />
       <button
         type="button"
