+++ client/src/components/Footer/style.scss
@@ -8,9 +8,6 @@
   bottom: 0px;
   right: 0px;
 
-  button {
-    display: flex;
-  }
   button:focus {
     outline: none;
   }
